% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveDescrStats.R
\name{SaveDescrStats}
\alias{SaveDescrStats}
\title{Saves all the components of the output of DescriptiveStats() to a specified folder}
\usage{
SaveDescrStats(
  DescriptiveStatsVar,
  path_stats,
  NumWidth = 1300,
  NumHeight = 800,
  CatWidth = 400,
  CatHeight = 400,
  TimeSeriesWidth = NULL,
  TimeSeriesHeight = NULL
)
}
\arguments{
\item{DefaultToExclude}{A Boolean. If TRUE, every non-ID and non-FK column will default to Include == FALSE, unless ExcludeIdentities or ExcludeForeignKeys is set to TRUE}
}
\description{
Gets the output of DescriptiveStats() and no matter what options were selected, saves in a specified folder all the valid Statistics
Saving the plots, one might also choose different sizes for the Categorical plots, Numerical ones, and Timeseries as well.
}
\examples{
DS <- mtcars \%>\% mutate(vs = as.factor(vs), am = as.factor(am), gear = as.factor(gear), carb = as.factor(carb)) \%>\% as_tibble()

#Creating the Variable which holds all the Matrices and Plots
MTCarsStats <- DS \%>\% DescriptiveStats(CalculateGraphs = TRUE, DependentVar = "mpg", IsTimeSeries = TRUE, GroupBy = "gear", CorrVarOrder = "PCA")

#Saving everything to a folder on our Hard Drive
MTCarsStats \%>\% SaveDescrStats("MTCarsFolder")
}
\keyword{Descriptive}
\keyword{Save}
\keyword{SaveDescrStats}
\keyword{Statistics}
