% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveDescrStats.R
\name{SaveDescrStats}
\alias{SaveDescrStats}
\title{Saves all the components of the output of DescriptiveStats() to a specified folder}
\usage{
SaveDescrStats(
  DescriptiveStatsVar,
  path_stats,
  NumWidth = 1300,
  NumHeight = 800,
  CatWidth = 1300,
  CatHeight = 800,
  TimeProgressionWidth = NULL,
  TimeProgressionHeight = NULL,
  TimeSeriesWidth = NULL,
  TimeSeriesHeight = NULL,
  CorrelationsWidth = NULL,
  CorrelationsHeight = NULL
)
}
\arguments{
\item{DescriptiveStatsVar}{A list. The output of DescriptiveStats() on a dataset}

\item{path_stats}{A string. The relative or full path of a folder (existing or to be created) where all the matrices and plots will be saved on}

\item{NumWidth}{An Integer. Width (in pixels) for Numerical Variables' plots}

\item{NumHeight}{An Integer. Height (in pixels) for Numerical Variables' plots}

\item{CatWidth}{An Integer. Width (in pixels) for Categorical Variables' plots}

\item{CatHeight}{An Integer. Height (in pixels) for Categorical Variables' plots}

\item{TimeProgressionWidth}{An Integer. Width (in pixels) for the Time Progression plot}

\item{TimeProgressionHeight}{An Integer. Height (in pixels) for the Time Progression plot}

\item{TimeSeriesWidth}{An Integer. Width (in pixels) for the Timeseries plot}

\item{TimeSeriesHeight}{An Integer. Height (in pixels) for the Timeseries plot}

\item{CorrelationsWidth}{An Integer. Width (in pixels) for the Correlation plots}

\item{CorrelationsHeight}{An Integer. Height (in pixels) for the Correlation plots}
}
\description{
Gets the output of DescriptiveStats() and no matter what options were selected, saves in a specified folder all the valid Statistics
Saving the plots, one might also choose different sizes for the Categorical plots, Numerical ones, and Timeseries as well.
}
\examples{
DS <- mtcars \%>\% mutate(vs = as.factor(vs), am = as.factor(am), gear = as.factor(gear), carb = as.factor(carb)) \%>\% as_tibble()

#Creating the Variable which holds all the Matrices and Plots
MTCarsStats <- DS \%>\% DescriptiveStats(CalculateGraphs = TRUE, DependentVar = "mpg", IsTimeSeries = TRUE, GroupBy = "gear", CorrVarOrder = "PCA")

#Saving everything to a folder on our Hard Drive
MTCarsStats \%>\% SaveDescrStats("MTCarsFolder")
}
\keyword{Descriptive}
\keyword{Save}
\keyword{SaveDescrStats}
\keyword{Statistics}
